Moves `<GameOverScreen>` to its own folder
And break it up into smaller components.
