+++ src/web-ui/screens/game-over/GameOverScreen.tsx
@@ -1,0 +1,26 @@
+import type { GameScreenProps } from '../../app';
+import { CtaButton } from '../../components/CtaButton';
+import './score.css';
+import HighScoresTable from './HighScoresTable';
+import PriceGuessHistoryBreakdown from './PriceGuessHistoryBreakdown';
+
+export const GameOverScreen = ({ gameCore }: GameScreenProps) => {
+  const score = gameCore.score.value;
+  const isPositive = score >= 0;
+
+  return (
+    <div className="game-over">
+      <div className={`score score--game-over ${isPositive ? 'score--positive' : 'score--negative'}`}>
+        <span>Final Score: {score}</span>
+        <span>{isPositive ? '🤑' : '💸'}</span>
+      </div>
+      <div className="game-over__content">
+        <PriceGuessHistoryBreakdown priceHistory={gameCore.priceHistory.value} />
+        <HighScoresTable />
+      </div>
+      <CtaButton variant="primary" className="game-over__restart" onClick={() => gameCore.restart()}>
+        Play Again
+      </CtaButton>
+    </div>
+  );
+};
