+++ src/web-ui/screens/game-over/highscore-form.css
