+++ src/web-ui/screens/game-over/score.css
@@ -1,0 +1,71 @@
+/* ========================================
+   Score Block (BEM Component) - Mobile First
+   ======================================== */
+.score {
+  font-weight: bold;
+  border-radius: 6px;
+  margin: 0;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  box-sizing: border-box;
+}
+
+/* Tablet and larger screens */
+@media (min-width: 768px) {
+  .score {
+    border-radius: 8px;
+  }
+}
+
+.score--game-loop {
+  font-size: 1rem;
+  padding: 0.5rem 0.75rem;
+  background-color: #fff8e1;
+  width: fit-content;
+  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.05);
+  color: #5d4037;
+  border: 1px solid #ffe082;
+  gap: 0.75rem;
+}
+
+/* Tablet and larger screens */
+@media (min-width: 768px) {
+  .score--game-loop {
+    font-size: 1.2rem;
+    padding: 0.6rem 0.9rem;
+    gap: 1rem;
+  }
+}
+
+.score--game-over {
+  font-size: 1.3rem;
+  padding: 1rem 1.25rem;
+  border-radius: 10px;
+  width: 100%;
+  max-width: 320px;
+  box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
+  transition: all 0.3s ease;
+}
+
+/* Tablet and larger screens */
+@media (min-width: 768px) {
+  .score--game-over {
+    font-size: 1.6rem;
+    padding: 1.25rem 1.75rem;
+    border-radius: 12px;
+    max-width: 400px;
+  }
+}
+
+.score--positive {
+  background-color: #d4edda;
+  color: #155724;
+  border: 2px solid #c3e6cb;
+}
+
+.score--negative {
+  background-color: #ffebee;
+  color: #721c24;
+  border: 2px solid #f5c6cb;
+}
